requirements.txt
transformers
torch --index-url https://download.pytorch.org/whl/cu121
torchvision --index-url https://download.pytorch.org/whl/cu121
torchaudio --index-url https://download.pytorch.org/whl/cu121
accelerate>=0.26.0
bitsandbytes
einops
peft
datasets
timm
llama-cpp-python
gguf
sentencepiece
protobuf
tensorboard
